Version 1.9.5
Add a tab "skip" list so that the toolbar button and keyboard shortcut don't load the listed tab and instead go to the previous one. Initially, the only tab on the list is the "Panorama View" tab of the Panorama Tab Groups extension. Ref. https://github.com/jscher2000/switch-to-previous-active-tab/issues/5